Komodo Robot Cup Detection
The general goal of this project is to implement "Detect and Reach"
algorithm, applied on Komodo robot (www.robotican.net) using the
increasing popular ROS (www.ros.org) — the robot's operating system. Our
algorithm uses Komodo's many sensors to scan an open space and spot a
disposable cup, reaching it while avoiding occurred obstacles, and
finally – grabbing the cup.
While Programming in Python and interfacing OpenCV's libraries with ROS,
this project provided us a hands-on and practical experience with issues
related to theoretical or applied topics in computational vision and/or
human perception.